fix test issue
