+++ tests/services/bookmarks-favorites.test.ts
@@ -1,5 +1,5 @@
 import { describe, it, expect, beforeEach } from 'vitest';
-import { setupAuth } from '../setup';
+import { setupAuth } from '../auth-helpers';
 import { mockBookmarks, mockFavorites } from '../mocks/mock-data';
 
 // Import the service functions we want to test
