+++ vitest.config.ts
@@ -6,7 +6,7 @@ export default defineConfig({
     include: ['tests/**/*.test.ts', 'src/router/**/*.test.ts'],
     globals: true,
     setupFiles: ['./tests/setup.ts'],
-    retry: 2,
+
     browser: {
       enabled: true,
       provider: playwright(),
